Refactor database, y README.md
