--- backend/app/database/__init__.py
@@ -1,1 +1,0 @@
-from .database import *
