+++ backend/pyproject.toml
@@ -13,6 +13,7 @@ uvicorn = "^0.32.0"
 sqlmodel = "^0.0.22"
 pymysql = "^1.1.1"
 python-dotenv = "^1.0.1"
+pytest = "^8.3.3"
 
 
 [build-system]
